packaging: Separate crop functionality and dependency
